[scrap] Support children when converting to SkSL to JavaScript
Include all child shaders (with ImageURLs and metadata) when
converting a shader from SkSL to JavaScript.

Bug: skia:14081
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/642896
Commit-Queue: Chris Mumford <cmumford@google.com>

+++ scrap/go/scrap/templates_test.go
@@ -16,7 +16,7 @@ func TestTemplateExpand_SVGToCPP_Success(t *testing.T) {
 		Type: SVG,
 		Body: "<svg> \n</svg>",
 	}
-	err = tmplMap[CPP][SVG].Execute(&b, body)
+	err = tmplMap[CPP][SVG].Execute(&b, scrapNode{Name: "Test", Scrap: body})
 	require.NoError(t, err)
 	expected := `void draw(SkCanvas* canvas) {
     const char * svg =
@@ -55,7 +55,7 @@ func TestTemplateExpand_SkSLToCPP_ResponseMatchesExpected(t *testing.T) {
 		Type: SKSL,
 		Body: "half4 main(in vec2 fragCoord ) {\n    return vec4( result, 1.0 );\n}",
 	}
-	err = tmplMap[CPP][SKSL].Execute(&b, body)
+	err = tmplMap[CPP][SKSL].Execute(&b, scrapNode{Name: "Test", Scrap: body})
 	require.NoError(t, err)
 	expected := `void draw(SkCanvas *canvas) {
     canvas->clear(SK_ColorBLACK);
@@ -106,7 +106,7 @@ func TestTemplateExpand_SkSLToCPPWithMetadata_ResponseMatchesExpected(t *testing
 		Body:         "// Inputs supplied by user:\nuniform float iSomeSlider;\n\nhalf4 main(in vec2 fragCoord ) {\n    return vec4( result, iSomeSlider );\n}",
 		SKSLMetaData: &SKSLMetaData{Uniforms: []float32{0.5}},
 	}
-	err = tmplMap[CPP][SKSL].Execute(&b, body)
+	err = tmplMap[CPP][SKSL].Execute(&b, scrapNode{Name: "Test", Scrap: body})
 	require.NoError(t, err)
 	expected := `void draw(SkCanvas *canvas) {
     canvas->clear(SK_ColorBLACK);
@@ -153,6 +153,7 @@ func TestTemplateExpand_SkSLToCPPWithMetadata_ResponseMatchesExpected(t *testing
 	require.Equal(t, expected, b.String())
 }
 
+// Test a simple scrap with no child shaders or custom uniform values.
 func TestTemplateExpand_SkSLToJavaScript_ResponseMatchesExpected(t *testing.T) {
 	tmplMap, err := loadTemplates()
 	require.NoError(t, err)
@@ -161,16 +162,19 @@ func TestTemplateExpand_SkSLToJavaScript_ResponseMatchesExpected(t *testing.T) {
 		Type: SKSL,
 		Body: "half4 main(in vec2 fragCoord ) {\n    return vec4( result, 1.0 );\n}",
 	}
-	err = tmplMap[JS][SKSL].Execute(&b, body)
+	// Create a scrap node with no name to test simpler variable names
+	err = tmplMap[JS][SKSL].Execute(&b, scrapNode{Scrap: body})
 	require.NoError(t, err)
-	expected := `const loadImage = fetch("https://shaders.skia.org/img/mandrill.png")
+	expected := `const shaderWidth = 512;
+const shaderHeight = 512;
+const loadImage1 = fetch("https://shaders.skia.org/img/mandrill.png")
   .then((response) => response.arrayBuffer());
 
-Promise.all([loadImage]).then((values) => {
-  const [imageData] = values;
-  const img = CanvasKit.MakeImageFromEncoded(imageData);
-  const imgShader = img.makeShaderCubic(
-    CanvasKit.TileMode.Clamp, CanvasKit.TileMode.Clamp, 1 / 3, 1 / 3);
+Promise.all([loadImage1]).then((values) => {
+  const [imageData1] = values;
+  const img1 = CanvasKit.MakeImageFromEncoded(imageData1);
+  const imgShader1 = img1.makeShaderCubic(
+    CanvasKit.TileMode.Clamp, CanvasKit.TileMode.Clamp, 1/3, 1/3);
 
   const surface = CanvasKit.MakeCanvasSurface(canvas.id);
   if (!surface) {
@@ -179,10 +183,10 @@ Promise.all([loadImage]).then((values) => {
   const skcanvas = surface.getCanvas();
   const paint = new CanvasKit.Paint();
   const startTimeMs = Date.now();
-  let mouseClickX = 250;
-  let mouseClickY = 250;
-  let mouseDragX = 250;
-  let mouseDragY = 250;
+  let mouseClickX = 0;
+  let mouseClickY = 0;
+  let mouseDragX = 0;
+  let mouseDragY = 0;
   let lastMousePressure = 0;
 
   const prog = ` + "`" + `
@@ -197,21 +201,23 @@ Promise.all([loadImage]).then((values) => {
         return vec4( result, 1.0 );
     }
     ` + "`" + `;
-
   const effect = CanvasKit.RuntimeEffect.Make(prog);
 
   function drawFrame(canvas) {
+    const iTime = (Date.now() - startTimeMs) / 1000;
     const uniforms = [
-      512, 512, 1,                                      // iResolution
-      (Date.now() - startTimeMs) / 1000,                // iTime
+      shaderWidth, shaderHeight, 1,                     // iResolution
+      iTime,                                            // iTime
       mouseDragX, mouseDragY, mouseClickX, mouseClickY, // iMouse
-      img.width(), img.height(), 1,                     // iImageResolution
-
+      img1.width(), img1.height(), 1,                   // iImageResolution
     ];
     const children = [
-      imgShader                                         // iImage1
+      imgShader1,                                       // iImage1
     ];
     const shader = effect.makeShaderWithChildren(uniforms, children);
+    if (!shader) {
+      throw "Could not make shader";
+    }
     paint.setShader(shader);
     skcanvas.drawPaint(paint);
     shader.delete();
@@ -237,25 +243,64 @@ Promise.all([loadImage]).then((values) => {
 	require.Equal(t, expected, b.String())
 }
 
-func TestTemplateExpand_SkSLWithMetadataToJavaScript_ResponseIncludesMetadata(t *testing.T) {
+// Test a more complex shader with custom uniforms and child shaders
+func TestTemplateExpand_SkSLWithChildNodesAndMetadata_ResponseIncludesMetadata(t *testing.T) {
 	tmplMap, err := loadTemplates()
 	require.NoError(t, err)
 	var b bytes.Buffer
-	body := ScrapBody{
-		Type:         SKSL,
-		Body:         "half4 main(in vec2 fragCoord ) {\n    uniform float  iSomeSlider;\n    return vec4( result, iSomeSlider );\n}",
-		SKSLMetaData: &SKSLMetaData{Uniforms: []float32{0.5}},
+	childA := scrapNode{
+		Name: "A",
+		Scrap: ScrapBody{
+			Type:         SKSL,
+			Body:         "uniform float  iValueA;\n\nhalf4 main(in vec2 fragCoord ) {\n    return vec4( result, iValueA );\n}",
+			SKSLMetaData: &SKSLMetaData{Uniforms: []float32{0.1}, ImageURL: "https://example.com/A.png"},
+		},
 	}
-	err = tmplMap[JS][SKSL].Execute(&b, body)
+	childB := scrapNode{
+		Name: "B",
+		Scrap: ScrapBody{
+			Type:         SKSL,
+			Body:         "uniform float  iValueB;\n\nhalf4 main(in vec2 fragCoord ) {\n    return vec4( result, iValueB );\n}",
+			SKSLMetaData: &SKSLMetaData{Uniforms: []float32{0.2}, ImageURL: "https://example.com/B.png"},
+		},
+	}
+	rootNode := scrapNode{
+		Name: "Root",
+		Scrap: ScrapBody{
+			Type: SKSL,
+			Body: `uniform float  iRootVal;
+
+half4 main(in vec2 fragCoord ) {
+    return vec4( result, iRootVal );
+}`,
+			SKSLMetaData: &SKSLMetaData{
+				Uniforms: []float32{0.3},
+				ImageURL: "https://example.com/A.png", // <- Second use of A.png
+				Children: []ChildShader{
+					{UniformName: "childA", ScrapHashOrName: "unused"},
+					{UniformName: "childB", ScrapHashOrName: "unused"},
+				},
+			},
+		},
+		Children: []scrapNode{childA, childB},
+	}
+	err = tmplMap[JS][SKSL].Execute(&b, rootNode)
 	require.NoError(t, err)
-	expected := `const loadImage = fetch("https://shaders.skia.org/img/mandrill.png")
+	expected := `const shaderWidth = 512;
+const shaderHeight = 512;
+const loadImage1 = fetch("https://example.com/A.png")
+  .then((response) => response.arrayBuffer());
+const loadImage2 = fetch("https://example.com/B.png")
   .then((response) => response.arrayBuffer());
 
-Promise.all([loadImage]).then((values) => {
-  const [imageData] = values;
-  const img = CanvasKit.MakeImageFromEncoded(imageData);
-  const imgShader = img.makeShaderCubic(
-    CanvasKit.TileMode.Clamp, CanvasKit.TileMode.Clamp, 1 / 3, 1 / 3);
+Promise.all([loadImage1, loadImage2]).then((values) => {
+  const [imageData1, imageData2] = values;
+  const img1 = CanvasKit.MakeImageFromEncoded(imageData1);
+  const imgShader1 = img1.makeShaderCubic(
+    CanvasKit.TileMode.Clamp, CanvasKit.TileMode.Clamp, 1/3, 1/3);
+  const img2 = CanvasKit.MakeImageFromEncoded(imageData2);
+  const imgShader2 = img2.makeShaderCubic(
+    CanvasKit.TileMode.Clamp, CanvasKit.TileMode.Clamp, 1/3, 1/3);
 
   const surface = CanvasKit.MakeCanvasSurface(canvas.id);
   if (!surface) {
@@ -264,13 +309,14 @@ Promise.all([loadImage]).then((values) => {
   const skcanvas = surface.getCanvas();
   const paint = new CanvasKit.Paint();
   const startTimeMs = Date.now();
-  let mouseClickX = 250;
-  let mouseClickY = 250;
-  let mouseDragX = 250;
-  let mouseDragY = 250;
+  let mouseClickX = 0;
+  let mouseClickY = 0;
+  let mouseDragX = 0;
+  let mouseDragY = 0;
   let lastMousePressure = 0;
 
-  const prog = ` + "`" + `
+  // Shader "A"
+  const progA = ` + "`" + `
     // Inputs supplied by shaders.skia.org:
     uniform float3 iResolution;      // Viewport resolution (pixels)
     uniform float  iTime;            // Shader playback time (s)
@@ -278,31 +324,104 @@ Promise.all([loadImage]).then((values) => {
     uniform float3 iImageResolution; // iImage1 resolution (pixels)
     uniform shader iImage1;          // An input image.
 
+    uniform float  iValueA;
+
     half4 main(in vec2 fragCoord ) {
-        uniform float  iSomeSlider;
-        return vec4( result, iSomeSlider );
+        return vec4( result, iValueA );
     }
     ` + "`" + `;
+  const effectA = CanvasKit.RuntimeEffect.Make(progA);
 
-  const effect = CanvasKit.RuntimeEffect.Make(prog);
+  // Shader "B"
+  const progB = ` + "`" + `
+    // Inputs supplied by shaders.skia.org:
+    uniform float3 iResolution;      // Viewport resolution (pixels)
+    uniform float  iTime;            // Shader playback time (s)
+    uniform float4 iMouse;           // Mouse drag pos=.xy Click pos=.zw (pixels)
+    uniform float3 iImageResolution; // iImage1 resolution (pixels)
+    uniform shader iImage1;          // An input image.
+
+    uniform float  iValueB;
+
+    half4 main(in vec2 fragCoord ) {
+        return vec4( result, iValueB );
+    }
+    ` + "`" + `;
+  const effectB = CanvasKit.RuntimeEffect.Make(progB);
+
+  // Shader "Root"
+  const progRoot = ` + "`" + `
+    // Inputs supplied by shaders.skia.org:
+    uniform float3 iResolution;      // Viewport resolution (pixels)
+    uniform float  iTime;            // Shader playback time (s)
+    uniform float4 iMouse;           // Mouse drag pos=.xy Click pos=.zw (pixels)
+    uniform float3 iImageResolution; // iImage1 resolution (pixels)
+    uniform shader iImage1;          // An input image.
+    uniform shader childA;
+    uniform shader childB;
+
+    uniform float  iRootVal;
+
+    half4 main(in vec2 fragCoord ) {
+        return vec4( result, iRootVal );
+    }
+    ` + "`" + `;
+  const effectRoot = CanvasKit.RuntimeEffect.Make(progRoot);
 
   function drawFrame(canvas) {
-    const uniforms = [
-      512, 512, 1,                                      // iResolution
-      (Date.now() - startTimeMs) / 1000,                // iTime
+    const iTime = (Date.now() - startTimeMs) / 1000;
+    const uniformsA = [
+      shaderWidth, shaderHeight, 1,                     // iResolution
+      iTime,                                            // iTime
       mouseDragX, mouseDragY, mouseClickX, mouseClickY, // iMouse
-      img.width(), img.height(), 1,                     // iImageResolution
-
+      img1.width(), img1.height(), 1,                   // iImageResolution
       // User supplied uniform values:
-      0.5
+      0.1
     ];
-    const children = [
-      imgShader                                         // iImage1
+    const childrenA = [
+      imgShader1,                                       // iImage1
     ];
-    const shader = effect.makeShaderWithChildren(uniforms, children);
-    paint.setShader(shader);
+    const shaderA = effectA.makeShaderWithChildren(uniformsA, childrenA);
+    if (!shaderA) {
+      throw "Could not make shaderA";
+    }
+    const uniformsB = [
+      shaderWidth, shaderHeight, 1,                     // iResolution
+      iTime,                                            // iTime
+      mouseDragX, mouseDragY, mouseClickX, mouseClickY, // iMouse
+      img2.width(), img2.height(), 1,                   // iImageResolution
+      // User supplied uniform values:
+      0.2
+    ];
+    const childrenB = [
+      imgShader2,                                       // iImage1
+    ];
+    const shaderB = effectB.makeShaderWithChildren(uniformsB, childrenB);
+    if (!shaderB) {
+      throw "Could not make shaderB";
+    }
+    const uniformsRoot = [
+      shaderWidth, shaderHeight, 1,                     // iResolution
+      iTime,                                            // iTime
+      mouseDragX, mouseDragY, mouseClickX, mouseClickY, // iMouse
+      img1.width(), img1.height(), 1,                   // iImageResolution
+      // User supplied uniform values:
+      0.3
+    ];
+    const childrenRoot = [
+      imgShader1,                                       // iImage1
+      shaderA,
+      shaderB,
+    ];
+    const shaderRoot = effectRoot.makeShaderWithChildren(uniformsRoot, childrenRoot);
+    if (!shaderRoot) {
+      throw "Could not make shaderRoot";
+    }
+    paint.setShader(shaderRoot);
     skcanvas.drawPaint(paint);
-    shader.delete();
+    shaderA.delete();
+    shaderB.delete();
+    shaderRoot.delete();
     surface.requestAnimationFrame(drawFrame);
   }
   surface.requestAnimationFrame(drawFrame);
@@ -351,7 +470,9 @@ func TestTemplateHelper_indentMultilineString_ReturnsExpectedIndent(t *testing.T
 func TestGetSkSLImageURL_WithValidPathsAndURLs_ReturnsObjectsShaderURL(t *testing.T) {
 	test := func(name, expected string, body ScrapBody) {
 		t.Run(name, func(t *testing.T) {
-			require.Equal(t, expected, getSkSLImageURL(body))
+			url, err := getSkSLImageURL(body)
+			assert.NoError(t, err)
+			require.Equal(t, expected, url)
 		})
 	}
 	test("DistNotModified", "https://shaders.skia.org/dist/soccer.png",
@@ -371,17 +492,14 @@ func TestGetSkSLImageURL_WithValidPathsAndURLs_ReturnsObjectsShaderURL(t *testin
 			SKSLMetaData: &SKSLMetaData{ImageURL: "https://example.com/my_image.png"}})
 }
 
-func TestGetSkSLImageURL_WithInvalidSkSLScrap_ReturnsDefaultShaderURL(t *testing.T) {
+func TestGetSkSLImageURL_NoScrapURL_ReturnsDefaultShaderURL(t *testing.T) {
 	test := func(name, expected string, body ScrapBody) {
 		t.Run(name, func(t *testing.T) {
-			require.Equal(t, expected, getSkSLImageURL(body))
+			url, err := getSkSLImageURL(body)
+			assert.NoError(t, err)
+			require.Equal(t, expected, url)
 		})
 	}
-	test("NotSkSL", "https://shaders.skia.org/img/mandrill.png",
-		ScrapBody{
-			Type:              Particle,
-			Body:              "",
-			ParticlesMetaData: &ParticlesMetaData{}})
 	test("NilSKSLMetaData", "https://shaders.skia.org/img/mandrill.png",
 		ScrapBody{
 			Type: SKSL,
@@ -392,18 +510,32 @@ func TestGetSkSLImageURL_WithInvalidSkSLScrap_ReturnsDefaultShaderURL(t *testing
 			SKSLMetaData: &SKSLMetaData{ImageURL: ""}})
 }
 
+func TestGetSkSLImageURL_WithInvalidSkSLScrap_ReturnsError(t *testing.T) {
+	test := func(name string, body ScrapBody) {
+		t.Run(name, func(t *testing.T) {
+			_, err := getSkSLImageURL(body)
+			require.Error(t, err)
+		})
+	}
+	test("NotSkSL",
+		ScrapBody{
+			Type:              Particle,
+			Body:              "",
+			ParticlesMetaData: &ParticlesMetaData{}})
+}
+
 func TestGetSkSLCustomUniforms_WithUniformValues_ReturnsCorrectStringForm(t *testing.T) {
 	test := func(name, expected string, body ScrapBody) {
 		t.Run(name, func(t *testing.T) {
 			require.Equal(t, expected, getSkSLCustomUniforms(body))
 		})
 	}
-	test("OneValue", "\n      // User supplied uniform values:\n      0.1",
+	test("OneValue", "      // User supplied uniform values:\n      0.1",
 		ScrapBody{
 			Type:         SKSL,
 			Body:         "",
 			SKSLMetaData: &SKSLMetaData{Uniforms: []float32{0.1}}})
-	test("TwoValues", "\n      // User supplied uniform values:\n      0.1, 0.2",
+	test("TwoValues", "      // User supplied uniform values:\n      0.1, 0.2",
 		ScrapBody{
 			Type:         SKSL,
 			Body:         "",
